move refreshing states to it's own method and call it initially
