+++ src/app/MSSFrontend.vcxproj.filters
@@ -227,4 +227,9 @@
       <Filter>Resource Files</Filter>
     </Image>
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="controllers\AppActionController\action_containers.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+  </ItemGroup>
 </Project>
